refactor: remove .component suffix from component name
