+++ src/components/NanoContract/EditAddressModal.js
